Added await to the Future.delayed and added check if two Todos are equal to replaceTodo

+++ riverpod_todo/lib/models/todo_notifier.dart
@@ -28,6 +28,7 @@ class TodoNotifier extends StateNotifier<List<Todo>> {
   }
 
   Future<void> replaceTodo(Todo oldTodo, Todo newTodo) async {
+    if (oldTodo == newTodo) return;
     await _repo.removeTodo(oldTodo);
     await addTodo(newTodo);
   }
